Make minutesToSeconds return long to avoid overflowing

+++ src/main/java/edu/hw1/Task1.java
@@ -9,7 +9,7 @@ public final class Task1 {
     }
 
     @SuppressWarnings("magicnumber")
-    public static int minutesToSeconds(@NotNull String string) {
+    public static long minutesToSeconds(@NotNull String string) {
         try (Scanner scanner = new Scanner(string).useDelimiter(":")) {
             if (!scanner.hasNextInt()) {
                 return -1;
@@ -19,12 +19,12 @@ public final class Task1 {
             if (!scanner.hasNextInt()) {
                 return -1;
             }
-            int seconds = scanner.nextInt();
+            long seconds = scanner.nextInt();
             if (minutes < 0 || seconds < 0 || seconds >= 60) {
                 return -1;
             }
 
-            seconds += minutes * 60;
+            seconds += (long) minutes * 60;
             return seconds;
         }
     }
